requirements.txt

Flask==3.0.3
pillow==10.3.0
pymongo==4.6.3
stegano==0.11.3
Werkzeug==3.0.2
gunicorn==20.1.0